Tiny Improvements on Methods - TextCrawler class
a. On the for loops, we can improve them by getting the **termination** condition once. In the loops edited above, they were constantly being called every time the loop ended.

b. On the boolean methods, we can just return the value of the statements instead of asking for their value first and then returning them.